Componentização para melhor organização e gerador de PDF
