+++ src/components/ClassRoom/ClassRoom.js
@@ -1,15 +1,12 @@
 import React, { Component } from "react"
-// import ReactDOMServer from "react-dom/server"
+import ReactDOMServer from "react-dom/server"
 import Fuse from "fuse.js"
 import Main from "../templates/Main"
 import API from "../../services/API"
-import CheckBox from "./form/CheckBox"
-import TextInput from "./form/TextInput"
-import Dropdown from "./form/Dropdown"
-import TextInputWithIcon from "./form/TextInputWithIcon"
-import Rows from "./table/Rows"
-import TableHeaders from "./table/TableHeaders"
-import SearchBar from "../TableHeader/SearchBar"
+import Form from "./Form"
+import Table from "./Table"
+import Filter from "./Filter"
+import TableOptions from "./TableOptions"
 const api = API("http://localhost:3001/classRooms")
 
 const headerProps = {
@@ -48,22 +45,11 @@ const initialState = {
 	},
 
 	saveButtonText: "Adicionar Sala",
-	showTableHeader: true,
+	showTableOptions: true,
 	showTable: true,
 	showForm: false,
 	showFilter: false
 }
-
-const tituloBlocoList = [
-	"Bloco A",
-	"Bloco B",
-	"Bloco C",
-	"Bloco D",
-	"Bloco E",
-	"Bloco F"
-]
-
-const numeroPisoList = [0, 1, 2]
 
 const thList = [
 	{
@@ -93,17 +79,6 @@ const thList = [
 	}
 ]
 
-const checkBoxStyle = {
-	margin: "10px"
-}
-
-const tableHeaderBtnStyle = {
-	color: "#FFF",
-	backgroundColor: "#2c3e50",
-	marginRight: "10px",
-	marginLeft: "10px",
-	marginBottom: "10px"
-}
 const fuseOptions = {
 	threshold: 0.6,
 	location: 0,
@@ -185,7 +160,7 @@ export default class ClassRoom extends Component {
 		this.setState({
 			showForm: !this.state.showForm,
 			showTable: !this.state.showTable,
-			showTableHeader: !this.state.showTableHeader
+			showTableOptions: !this.state.showTableOptions
 		})
 	}
 
@@ -314,184 +289,71 @@ export default class ClassRoom extends Component {
 			this.setState({ list })
 		}
 	}
-	renderTableHeader = () => {
-		if (this.state.showTableHeader) {
+
+	generatePDF = () => {
+		const html = ReactDOMServer.renderToStaticMarkup(
+			<Table
+				thList={thList}
+				listOrder={this.state.listOrder}
+				onClick={this.listSort}
+				list={this.state.list}
+				remove={this.remove}
+				load={this.load}
+			/>
+		)
+		api.fetchAndGetList(html)
+	}
+
+	renderTableOptions = () => {
+		if (this.state.showTableOptions) {
 			return (
-				<div className="row">
-					<div className="col">
-						<button
-							style={tableHeaderBtnStyle}
-							className="btn float-left"
-							onClick={() => this.formToggle()}
-						>
-							<i className="fa fa-plus fa-fw" />
-							Adicionar Sala
-						</button>
-						<button
-							style={tableHeaderBtnStyle}
-							className="btn float-left"
-							onClick={() => this.filterToggle()}
-						>
-							<i className="fa fa-filter fa-fw" />
-							Filtro
-						</button>
-					</div>
-					<div className="col">
-						<SearchBar
-							value={this.state.searchQuery}
-							placeholder="Pesquisa"
-							onChange={value => this.updateSearchQuery(value)}
-						/>
-					</div>
-				</div>
+				<TableOptions
+					formToggle={this.formToggle}
+					filterToggle={this.filterToggle}
+					searchQuery={this.state.searchQuery}
+					searchOnChange={this.updateSearchQuery}
+					generatePDF={this.generatePDF}
+				/>
 			)
 		}
 	}
 
 	renderFilter = () => {
-		if (this.state.showFilter && this.state.showTableHeader) {
+		if (this.state.showFilter && this.state.showTableOptions) {
 			return (
-				<div className="form-group">
-					<div style={checkBoxStyle} className="form-row">
-						{this.renderCheckBoxes(tituloBlocoList, "titulo_bloco")}
-					</div>
-					<div style={checkBoxStyle} className="form-row">
-						{this.renderCheckBoxes(numeroPisoList, "numero_piso")}
-						<div className="col" />
-						<div className="col" />
-						<div className="col" />
-					</div>
-				</div>
+				<Filter
+					isChecked={this.state.isChecked}
+					onChange={this.checkBoxToggle}
+				/>
 			)
 		}
-	}
-
-	renderCheckBoxes = (values, field) => {
-		let complement = ""
-		if (field === "numero_piso") complement = "Piso "
-		return values.map(item => {
-			return (
-				<div className="col">
-					<CheckBox
-						label={`${complement}${item}`}
-						value={item}
-						onChange={event => this.checkBoxToggle(event, field)}
-						isChecked={this.state.isChecked[field][item]}
-					/>
-				</div>
-			)
-		})
 	}
 
 	renderTable = () => {
 		if (this.state.showTable) {
 			return (
-				<div>
-					<table className="table mt-4">
-						<thead>
-							<tr>
-								<TableHeaders
-									values={thList}
-									listOrder={this.state.listOrder}
-									onClick={this.listSort}
-								/>
-							</tr>
-						</thead>
-						<tbody>
-							<Rows
-								values={this.state.list}
-								remove={this.remove}
-								load={this.load}
-							/>
-						</tbody>
-					</table>
-				</div>
+				<Table
+					thList={thList}
+					listOrder={this.state.listOrder}
+					onClick={this.listSort}
+					list={this.state.list}
+					remove={this.remove}
+					load={this.load}
+				/>
 			)
 		}
 	}
 
-	renderTh = () => {
-		let icon
-		return thList.map(item => {
-			if (item.showSort) {
-				icon =
-					this.state.listOrder === "increasing" ? "fa-sort-up" : "fa-sort-down"
-			} else {
-				icon = "fa-sort text-muted"
-			}
-			return (
-				<th onClick={() => this.listSort(item.id)}>
-					<span>
-						{item.label}
-						<i className={`fa ${icon} fa-fw`} />
-					</span>
-				</th>
-			)
-		})
-	}
-
 	renderForm = () => {
 		if (this.state.showForm) {
-			const {
-				titulo_bloco,
-				numero_piso,
-				codigo_sala,
-				titulo_sala
-			} = this.state.classRoom
 			return (
-				<form className="form">
-					<div className="row">
-						<Dropdown
-							values={tituloBlocoList}
-							label="Titulo do Bloco"
-							name="titulo_bloco"
-							value={titulo_bloco}
-							onChange={async event => await this.updateField(event)}
-						/>
-
-						<Dropdown
-							values={numeroPisoList}
-							label="Numero do Piso"
-							name="numero_piso"
-							value={numero_piso}
-							onChange={async event => await this.updateField(event)}
-						/>
-
-						<TextInputWithIcon
-							values={[titulo_bloco[6], numero_piso]}
-							label="Codigo da Sala"
-							name="codigo_sala"
-							value={codigo_sala}
-							onChange={async event => await this.updateField(event)}
-							maxLength={3}
-						/>
-
-						<TextInput
-							label="Titulo da Sala"
-							name="titulo_sala"
-							value={titulo_sala}
-							onChange={async event => await this.updateField(event)}
-						/>
-					</div>
-					<div className="row">
-						<div className="col-12 d-flex justify-content-end">
-							<button
-								type="button"
-								onClick={() => this.save(this.state.classRoom)}
-								className="btn btn-primary"
-							>
-								{this.state.saveButtonText}
-							</button>
-							<button
-								type="button"
-								className="btn btn-secondary ml-2"
-								onClick={() => this.clear()}
-							>
-								Cancelar
-							</button>
-						</div>
-					</div>
-				</form>
+				<Form
+					classRoom={this.state.classRoom}
+					updateField={this.updateField}
+					save={this.save}
+					clear={this.clear}
+					saveButtonText={this.state.saveButtonText}
+				/>
 			)
 		}
 	}
@@ -499,7 +361,7 @@ export default class ClassRoom extends Component {
 	render() {
 		return (
 			<Main {...headerProps}>
-				{this.renderTableHeader()}
+				{this.renderTableOptions()}
 				{this.renderFilter()}
 				{this.renderForm()}
 				{this.renderTable()}
