--- src/components/ClassRoom/form/CheckBox.jsx
@@ -1,17 +1,0 @@
-import React from "react"
-
-export default props => {
-	const { label, onChange, isChecked, value } = props
-	return (
-		<div className="form-check form-check-inline">
-			<input
-				type="checkbox"
-				value={value}
-				onChange={onChange}
-				checked={isChecked}
-				className="form-check-input"
-			/>
-			<label className="form-check-label">{label}</label>
-		</div>
-	)
-}
